requirements.txt
asgiref==3.6.0
astroid==2.11.7
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==6.3.2
cryptography==39.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-admin-thumbnails==0.2.6
django-allauth==0.52.0
django-auto-prefetch==1.5.0
django-countries==7.5.1
django-db-connection-pool==1.2.2
django-debug-toolbar==3.8.1
django-redis==5.2.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
Faker==17.6.0
flake8==4.0.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
ipython==8.0.1
isort==5.10.1
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.13.0
pytest==7.2.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.6.0
pytest-mock==3.7.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pyuca==1.2
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
